优化 ubuntu-gpu-pv.ps1 和 win-gpu-pv.ps1 脚本，添加条件检查以确保在移除 GPU 分区适配器之前存在该适配器，提升脚本的健壮性。

+++ ubuntu-gpu-pv.ps1
@@ -36,7 +36,7 @@ Write-Host "Stopping VM"
 $vmobject | Stop-VM
 
 Write-Host "Configuring GPU-PV for VM"
-$vmobject | Remove-VMGpuPartitionAdapter
+if ($vmobject | Get-VMGpuPartitionAdapter) { $vmobject | Remove-VMGpuPartitionAdapter }
 $vmobject | Add-VMGpuPartitionAdapter -InstancePath "$path"
 $vmobject | Set-VMGpuPartitionAdapter  `
 	-MinPartitionVRAM $targetGpu.MinPartitionVRAM `
